+++ CHANGELOG.md
@@ -6,6 +6,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [0.26.0] - 2023-08-19
+### Changed
+- [@mjkuranda](https://github.com/mjkuranda): Destroying an asteroid gives 2-4 new missiles.
+
 ### Fixed
 - [@mjkuranda](https://github.com/mjkuranda): Game animations while switching any renderer.